requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
blis==1.3.3
build==1.4.0
cachetools==6.2.4
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docx2pdf==0.1.8
docxtpl==0.20.2
drf-spectacular==0.29.0
durationpy==0.10
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
gensim==4.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-genai==1.59.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflection==0.5.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.13.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langsmith==0.6.4
lark==1.3.1
libclang==18.1.1
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
notebook==7.5.2
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
posthog==5.4.0
preshed==3.0.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pypdf==6.6.0
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-discovery==1.1.1
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
pywin32==311
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==2.1.0
sentence-transformers==5.2.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
sqlparse==0.5.5
srsly==2.5.2
stack-data==0.6.3
streamlit==1.53.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
terminado==0.18.1
tf_keras==2.20.1
thinc==8.3.10
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
virtualenv==21.1.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.5
widgetsnbextension==4.0.15
wrapt==2.0.1
yarl==1.22.0
yt-dlp==2025.12.8
zipp==3.23.0
zstandard==0.25.0
